Move some utils to nengo-extras
Added to nengo-extras in nengo/nengo-extras#29.

- Move graphviz functions
- Move functions for estimating firing rates
- Move `probe_all` function
